refactor(Folder): Organizing repeated functions just in one file
