+++ src/utils/Cargo.toml
@@ -15,6 +15,13 @@ once_cell = "1.13"
 dirs = "4.0.0"
 clap = { version = "3.2", features = ["derive"] }
 anyhow = "1.0"
+base64 = "0.13.1"
+tokio = { version = "1", features = ["full"] }
+tokio-util = { version = "0.7", features = ["full"] }
+image = "0.24.5"
+futures = "0.3"
+mime = "0.3.16"
+log = "0.4.17"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 cocoa = "0.24.1"
